Changing ServerMain to make it non-static
I used a project I am making using a video tutorial online as a template
to add keyboard functionality. I realized that the static way i was
doing this was not ideal and copied the non-static implementation that
implemented runnable. Now we have keyboard input with the up key and a
skeleton gui.